requirements.txt
requests
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
langchain
streamlit
plotly
dash
flask